atualizando um item ja existente

+++ js/main2.js
@@ -11,13 +11,26 @@ form.addEventListener("submit", (event) => {
     event.preventDefault(); //interromper o 'enviar' informações para mesma tela no "submit"
     const nome = event.target.elements["nome"]; //colocando o nome digitado na variavel 'nome'
     const quantidade = event.target.elements["quantidade"]; //colocando a quantidade digitado na variável
+
+    const existe = itens.find(element => element.nome === nome.value);
+
     const itemAtual = {
         "nome": nome.value,
         "quantidade": quantidade.value
     } //criando um objeto para salvar no localStorage
-    criaElemento(itemAtual); //chamar a função para adicionar item na lista
-    itens.push(itemAtual); // adicionado o objeto no array []
-    localStorage.setItem("itens", JSON.stringify(itens)); //incluindo o objeto dentro do localStorage dentro da Key 'itens'
+
+    if (existe) {
+        console.log("Item Repetido");
+        itemAtual.id = existe.id;
+        atualizaQtd(itemAtual);
+    }
+    else{
+        console.log("Item novo!");
+        itemAtual.id = itens.length;
+        criaElemento(itemAtual); //chamar a função para adicionar item na lista
+        itens.push(itemAtual); // adicionado o objeto no array []
+        localStorage.setItem("itens", JSON.stringify(itens)); //incluindo o objeto dentro do localStorage dentro da Key 'itens'
+    }
     nome.value = ""; //zerar campo
     quantidade.value = ""; //zerar campo
 })
@@ -27,8 +40,12 @@ function criaElemento(item) {
     novoItem.classList.add("item"); // adicionando a classe "item" no elemento 'novoItem'
     const qtdItem = document.createElement('strong'); //criando outro elemento, dessa vez com a tag <strong>
     qtdItem.innerHTML = item.quantidade; // inserindo dentro do elemento 'qtdItem' a quantidade do objeto enviado como parâmetro
+    qtdItem.dataset.id = item.id;
     novoItem.appendChild(qtdItem); //inputando o elemento 'qtdItem' para dntro do elemento 'novoItem'
     novoItem.innerHTML += item.nome; // adicionando dentro do elemento 'novoItem' o 'nome' objeto enviado como parâmetro
     lista.appendChild(novoItem); //inputando o elemento 'novoItem' para dntro da lista
 }
 
+function atualizaQtd(item){
+    document.querySelector("[data-id='"+item.id+"']").innerHTML = item.quantidade;
+}
